+++ src/api/mod.rs
@@ -1,5 +1,4 @@
 pub mod api_utils;
-pub mod api_model;
 pub mod main_api;
 mod download_api;
 mod v1_api;
@@ -8,3 +7,5 @@ mod m3u_api;
 mod xmltv_api;
 mod scheduler;
 mod web_index;
+
+pub(crate) mod model;