+++ src/api/scheduler.rs
@@ -4,7 +4,7 @@ use actix_web::web::Data;
 use chrono::{DateTime, FixedOffset, Local};
 use cron::Schedule;
 use log::error;
-use crate::api::api_model::AppState;
+use crate::api::model::app_state::AppState;
 use crate::exit;
 use crate::processing::playlist_processor::exec_processing;
 
@@ -31,7 +31,7 @@ pub async fn start_scheduler(expression: &str, data: Data<AppState>) -> ! {
             loop {
                 let mut upcoming = schedule.upcoming(offset).take(1);
                 if let Some(datetime) = upcoming.next() {
-                    actix_rt::time::sleep_until(actix_rt::time::Instant::from(datetime_to_instant(datetime))).await;
+                    actix_web::rt::time::sleep_until(actix_rt::time::Instant::from(datetime_to_instant(datetime))).await;
                     exec_processing(data.config.clone(), data.targets.clone()).await;
                 }
             }
@@ -63,7 +63,7 @@ mod tests {
                 loop {
                     let mut upcoming = schedule.upcoming(offset).take(1);
                     if let Some(datetime) = upcoming.next() {
-                        actix_rt::time::sleep_until(actix_rt::time::Instant::from(datetime_to_instant(datetime))).await;
+                        actix_web::rt::time::sleep_until(actix_rt::time::Instant::from(datetime_to_instant(datetime))).await;
                         run_me();
                     }
                     if runs.load(Ordering::Relaxed) == 6 {