+++ src/repository/epg_repository.rs
@@ -1,8 +1,9 @@
 use std::fs::File;
 use std::io::{Cursor, Write};
 use std::path::{Path};
-use log::{debug, log_enabled, Level};
+use log::{Level};
 use quick_xml::{Writer};
+use crate::debug_if_enabled;
 use crate::m3u_filter_error::{M3uFilterError, M3uFilterErrorKind};
 use crate::model::config::{Config, ConfigTarget, TargetOutput};
 use crate::model::config::TargetType;
@@ -24,9 +25,7 @@ fn epg_write_file(target: &ConfigTarget, epg: &Epg, path: &Path) -> Result<(), M
                     }
                     match epg_file.write_all(&result) {
                         Ok(()) => {
-                            if log_enabled!(Level::Debug) {
-                                debug!("Epg for target {} written to {}", target.name, path.to_str().unwrap_or("?"));
-                            }
+                           debug_if_enabled!("Epg for target {} written to {}", target.name, path.to_str().unwrap_or("?"));
                         }
                         Err(err) => return Err(M3uFilterError::new(
                             M3uFilterErrorKind::Notify, format!("failed to write epg: {} - {}", path.to_str().unwrap_or("?"), err))),
@@ -52,18 +51,14 @@ pub fn epg_write(target: &ConfigTarget, cfg: &Config, target_path: &Path, epg: O
                         format!("write epg for target {} failed: No filename set", target.name)));
                 }
                 let path = m3u_get_epg_file_path(target_path);
-                if log_enabled!(Level::Debug) {
-                    debug!("writing m3u epg to {}", path.to_str().unwrap_or("?"));
-                }
+                debug_if_enabled!("writing m3u epg to {}", path.to_str().unwrap_or("?"));
                 epg_write_file(target, epg_data, &path)?;
             }
             TargetType::Xtream => {
                 match xtream_get_storage_path(cfg, &target.name) {
                     Some(path) => {
                         let epg_path = xtream_get_epg_file_path(&path);
-                        if log_enabled!(Level::Debug) {
-                            debug!("writing xtream epg to {}", epg_path.to_str().unwrap_or("?"));
-                        }
+                        debug_if_enabled!("writing xtream epg to {}", epg_path.to_str().unwrap_or("?"));
                         epg_write_file(target, epg_data, &epg_path)?;
                     }
                     None => return Err(M3uFilterError::new(