+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
-# Changelog
 # 2.0.11 (2024-12-xx)
 - no error log for xtream api when content is skipped with options `xtream_skip_[live|vod|series]`
+- added live channel connection sharing in reverse proxy mode. To activate set `share_live_streams` in target options. 
 
 # 2.0.10 (2024-12-03)
 - added Target Output Option `m3u_include_type_in_url`, default false. This adds `live`, `movie`, `series` to the url of the stream in reverse proxy mode.